feat: 영수증 OCR 기능 구현

+++ src/main/java/com/example/hackathon/receipt/repository/ReceiptRepository.java
@@ -1,8 +1,8 @@
 // src/main/java/com/example/hackathon/receipt/repository/ReceiptRepository.java
 package com.example.hackathon.receipt.repository;
 
-import com.example.hackathon.receipt.entity.Receipt;
 import com.example.hackathon.receipt.OcrStatus;
+import com.example.hackathon.receipt.entity.Receipt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -12,7 +12,8 @@ import java.util.Optional;
 @Repository
 public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
 
-    Optional<Receipt> findByIdAndUser_Id(Long id, Integer userId);
+    // userId를 Long으로 통일
+    Optional<Receipt> findByIdAndUser_Id(Long id, Long userId);
 
     List<Receipt> findByUserMission_IdOrderByIdDesc(Long userMissionId);
